+++ apps/playground/vite.config.ts
@@ -1,9 +1,9 @@
-import vue from "@vitejs/plugin-vue";
 import { defineConfig } from "vite";
+import { VITE_BASE_CONFIG } from "./node_modules/@sit-onyx/shared/vite.config.base";
 
 // https://vitejs.dev/config
 export default defineConfig({
-  plugins: [vue()],
+  ...VITE_BASE_CONFIG,
   optimizeDeps: {
     // needed in order for the web workers of "@vue/repl" to work
     exclude: ["@vue/repl"],