+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-beta.58
+
+### Patch Changes
+
+- Updated dependencies [a9d89f6]
+  - sit-onyx@1.0.0-beta.52
+
+## 1.0.0-beta.57
+
+### Patch Changes
+
+- Updated dependencies [6fe0527]
+  - sit-onyx@1.0.0-beta.51
+
+## 1.0.0-beta.56
+
+### Patch Changes
+
+- Updated dependencies [beca92e]
+  - sit-onyx@1.0.0-beta.50
+
 ## 1.0.0-beta.55
 
 ### Patch Changes